refactor(files): update delete endpoint to use soft delete with deleted_at

+++ app/api/V1/file.py
@@ -1,14 +1,15 @@
 import uuid
-from datetime import datetime, timedelta
+from datetime import datetime, timedelta, timezone
 from fastapi import APIRouter, Depends, HTTPException, Request, Response, UploadFile
 from sqlalchemy.orm import Session
 from app.core.database import get_db
 from app.core.security import get_current_user
 from app.models.file import File
+from app.models.folder import Folder
 from app.models.user import User
-from app.schemas.file import FileRename, ShareFileSchema
+from app.schemas.file import FileRename, FileResponse, MoveFileSchema, ShareFileSchema
 from app.models.fileshare import FileShare
-from app.services.file_service import delete_file_from_b2, download_file_from_b2, upload_file_to_b2
+from app.services.file_service import delete_file_from_b2, download_file_from_b2, get_accesible_file, upload_file_to_b2
 
 router = APIRouter(prefix="/file", tags=["file"])
 # Subir archivo
@@ -23,16 +24,42 @@ def upload_file(folder_id: str, file: UploadFile, db: Session = Depends(get_db),
 @router.get("/{file_id}/download")
 def download_file(file_id: str, db: Session = Depends(get_db), current_user=Depends(get_current_user)):
 
-    ## Solo puede descargar el archivo el owner, modificar para shares
     ## Crear funcion get_accesible_file, para evitar reutilizar codigo xd
-    print(file_id)
-    file = db.query(File).filter(File.id == file_id, File.user_id == current_user.id).first()
+
+    file = get_accesible_file(db, file_id, current_user.id)
+
     if not file:
-        raise HTTPException(status_code=404, detail="Archivo no encontrado")
-    print(file.path)
+            raise HTTPException(status_code=404, detail="Archivo no encontrado")
+    
     download_url = download_file_from_b2(file.path)
     return {"download_url": download_url}
+
+
+@router.post("/{file_id}/restore")
+def restore_file(file_id: uuid.UUID, db: Session = Depends(get_db), current_user = Depends(get_current_user)):
+    file = db.query(File).filter(File.id == file_id, File.user_id == current_user.id).first()
+
+
+
+    if not file:
+        raise HTTPException(
+            status_code=404,
+            detail="File not found"
+            )
     
+    if file.deleted_at is None:
+        raise HTTPException(
+            status_code=400,
+            detail="file is not deleted"
+        )    
+    
+    file.deleted_at = None
+
+    db.commit()
+
+
+
+
 
 # Borrar archivo
 @router.delete("/{file_id}")
@@ -41,31 +68,51 @@ def delete_file(file_id: str, db: Session = Depends(get_db), current_user=Depend
 
     if not file:
         raise HTTPException(status_code=404, detail="Archivo no encontrado")
-    
-    db.delete(file)
+
+    if file.deleted_at is not None:
+            delete_file_from_b2(file.path)
+            db.delete(file)
+            db.commit()
+
+            return {"message": "Archivo eliminado exitosamente"}
+
+    file.deleted_at = datetime.now(timezone.utc)
     db.commit()
 
-    delete_file_from_b2(file.path)  
-    return {"message": "Archivo eliminado exitosamente"}
+    return {"message": "Archivo puesto en la papelera"}
 
 # Mover archivo a otra carpeta
 @router.put("/{file_id}/move")
-def move_file(file_id: uuid.UUID, new_folder_id: uuid.UUID, db: Session = Depends(get_db), current_user=Depends(get_current_user)):
+def move_file(file_id: uuid.UUID, move_data: MoveFileSchema, db: Session = Depends(get_db), current_user=Depends(get_current_user)):
     # Implementar logica mover archivo
 
     file = db.query(File).filter(File.id == file_id, File.user_id==current_user.id).first()
-
+    
     if not file:
         raise HTTPException(status_code=404, detail="Archivo no encontrado")
 
-    file.folder_id = new_folder_id
+    if file.deleted_at is not None:
+        raise HTTPException(
+            status_code=400,
+            detail= "Cannot move a deleted file"
+            )
+
+    folder = db.query(Folder).filter(Folder.id == move_data.new_folder_id, Folder.user_id == current_user.id).first()
+
+    if not folder:
+        raise HTTPException(
+            status_code=404,
+            detail="folder not found"
+            )
+
+    file.folder_id = move_data.new_folder_id
 
     db.commit()
 
     return {"message": "file moved successfully"}
 
 # Renombrar archivo
-@router.put("/{file_id}/rename")
+@router.put("/{file_id}/rename", response_model=FileResponse)
 def rename_file(file_id: uuid.UUID, fileResponse: FileRename, db: Session = Depends(get_db), current_user=Depends(get_current_user)):
     #logica renombrar archivo
 
@@ -73,23 +120,14 @@ def rename_file(file_id: uuid.UUID, fileResponse: FileRename, db: Session = Depe
     if not file:
         raise HTTPException(status_code=404, detail="file not found")
 
-
-    ## Crear fileResponse 
-    new_file = File(
-        folder_id = file.folder_id,
-        name = file.name,
-        size = file.size,
-        mime_type = file.mime_type,
-    )
-
     file.name = fileResponse.name 
     db.commit()
 
-    return {"message": "File renamed successfully", "file": new_file, "file2xd": file}
+    return file
 
 
 # Obtener detalles de un archivo
-@router.get("/{file_id}")
+@router.get("/{file_id}", response_model=FileResponse)
 def get_file(file_id: uuid.UUID, db: Session = Depends(get_db), current_user = Depends(get_current_user)):
     file = db.query(File).filter(File.id== file_id).first()
     
@@ -97,15 +135,14 @@ def get_file(file_id: uuid.UUID, db: Session = Depends(get_db), current_user = D
         raise HTTPException(status_code=404, detail="file not found")
     
 
-    new_file = new_file = File(
-        folder_id = file.folder_id,
-        name = file.name,
-        size = file.size,
-        mime_type = file.mime_type,
-    )
-
     if file.user_id == current_user.id:
-        return {"file": new_file}
+        return file
+    
+    if file.deleted_at is not None:
+        raise HTTPException(
+            status_code=404,
+            detail="file not found"
+        )
     
     share = db.query(FileShare).filter(
         FileShare.file_id == file_id,
@@ -115,17 +152,24 @@ def get_file(file_id: uuid.UUID, db: Session = Depends(get_db), current_user = D
     if not share:
         raise HTTPException(
             status_code=403,
-            detail="Acces denied"
+            detail="Access denied"
         )
     
-    if share.expires_at and share.expires_at < datetime.utcnow(): 
+    if share.expires_at and share.expires_at < datetime.now(timezone.utc): 
         raise HTTPException(
             status_code=403,
             detail="Share expired"
         )
 
-    return {"file": new_file}
+    return file
     
+
+# Ver archivos en la papelera implementar despues   
+@router.get("/trash")
+def get_file_deleted(db: Session = Depends(get_db), current_user = Depends(get_current_user)):
+    "Logica para mostrar todos los archivos eliminados"
+
+
 
 
 # Compartir archivo con otro usuario
@@ -141,7 +185,7 @@ def share_file(file_id : uuid.UUID,share_data: ShareFileSchema, db: Session = De
             share_data.expires_in_days = 30
 
 
-        expires_at = datetime.utcnow() + timedelta(
+        expires_at = datetime.now(timezone.utc) + timedelta(
         days= share_data.expires_in_days or 0,
         hours= share_data.expires_in_hours or 0,
         minutes= share_data.expires_in_minutes or 0,
@@ -159,7 +203,7 @@ def share_file(file_id : uuid.UUID,share_data: ShareFileSchema, db: Session = De
         )
 
     file = db.query(File).filter(File.id == file_id, File.
-    user_id == current_user.id, File.is_deleted == False).first()
+    user_id == current_user.id, File.deleted_at == None).first()
     if not file:
         raise HTTPException(status_code=404, detail="File not found")
 
@@ -168,11 +212,11 @@ def share_file(file_id : uuid.UUID,share_data: ShareFileSchema, db: Session = De
 
     if existing_share:
         existing_share.expires_at = expires_at
-        existing_share.shared_at = datetime.utcnow()
+        existing_share.shared_at = datetime.now(timezone.utc)
         db.commit()
 
         return {
-        "message": "File shared, succesfully",
+        "message": "File shared succesfully",
         "share_id": str(existing_share.id),
         "expires_at": expires_at
     }
@@ -184,9 +228,9 @@ def share_file(file_id : uuid.UUID,share_data: ShareFileSchema, db: Session = De
         id= uuid.uuid4(),
         file_id= file_id,
         shared_with_user_id= user_to_share.id,
-        created_at= datetime.utcnow(),
+        created_at= datetime.now(timezone.utc),
         expires_at=expires_at,
-        shared_at= datetime.utcnow(),
+        shared_at= datetime.now(timezone.utc),
     )
 
     db.add(file_share)
